+++ lib/state/lruqueue.py
@@ -1,5 +1,5 @@
 from collections import OrderedDict
-from typing import Any
+from typing import Any, Optional
 
 
 class LRUQueue: